+++ app/src/test/java/org/simple/clinic/contactpatient/SetAppointmentReminderUiRendererTest.kt
@@ -43,6 +43,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).enablePreviousReminderDateStepper()
     verify(ui).enableNextReminderDateStepper()
@@ -67,6 +68,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).renderSelectedAppointmentDate(reminderPeriod, selectedReminderDate.scheduledFor)
     verify(ui).enableNextReminderDateStepper()
@@ -91,6 +93,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).renderSelectedAppointmentDate(reminderPeriod, selectedReminderDate.scheduledFor)
     verify(ui).enableNextReminderDateStepper()
@@ -115,6 +118,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).renderSelectedAppointmentDate(reminderPeriod, selectedReminderDate.scheduledFor)
     verify(ui).enablePreviousReminderDateStepper()
@@ -139,6 +143,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).renderSelectedAppointmentDate(reminderPeriod, selectedReminderDate.scheduledFor)
     verify(ui).enablePreviousReminderDateStepper()
@@ -163,6 +168,7 @@ class SetAppointmentReminderUiRendererTest {
 
     // then
     verify(ui).hideProgress()
+    verify(ui).hideCallResult()
     verify(ui).switchToSetAppointmentReminderView()
     verify(ui).renderSelectedAppointmentDate(reminderPeriod, selectedReminderDate.scheduledFor)
     verify(ui).enablePreviousReminderDateStepper()